ENG-578: Move base64 into crypto
