+++ fendermint/app/src/cmd/key.rs
@@ -3,7 +3,7 @@
 
 use anyhow::{anyhow, Context};
 use fendermint_app_options::key::KeyShowPeerIdArgs;
-use fendermint_crypto::{PublicKey, SecretKey};
+use fendermint_crypto::{from_b64, to_b64, PublicKey, SecretKey};
 use fendermint_vm_actor_interface::eam::EthAddress;
 use fvm_shared::address::Address;
 use rand_chacha::{rand_core::SeedableRng, ChaCha20Rng};
@@ -11,7 +11,6 @@ use serde_json::json;
 use std::path::{Path, PathBuf};
 use tendermint_config::NodeKey;
 
-use super::{from_b64, to_b64};
 use crate::{
     cmd,
     options::key::{
